+++ packages/pigeon/pigeons/message.dart
@@ -2,7 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:pigeon/pigeon_lib.dart';
+// This file is an example pigeon file that is used in compilation, unit, mock
+// handler, and e2e tests.
+
+import 'package:pigeon/pigeon.dart';
 
 class SearchRequest {
   String query;
@@ -15,7 +18,7 @@ class SearchReply {
   String error;
 }
 
-@HostApi()
+@HostApi(dartHostTestHandler: 'TestHostApi')
 abstract class Api {
   SearchReply search(SearchRequest request);
 }